% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{adf.example}
\alias{adf.example}
\title{An example of an amigaDisk object}
\format{An S4 \code{\link{amigaDisk-class}} object.}
\description{
An example of an \code{\link{amigaDisk-class}} object.
}
\details{
An \code{\link{amigaDisk-class}} object that represents an
'Old File System' formatted and bootable disk. It is used in
multiple examples of this package. It contains a directory
structure and some files that can be accessed using this package.
The content of this example might change in future versions.
}
\examples{
data("adf.example")
}
