% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adfExplorer-package.R
\docType{package}
\name{adfExplorer-package}
\alias{adfExplorer}
\alias{adfExplorer-package}
\title{adfExplorer: Access and Manipulate Amiga Disk Files}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Amiga Disk Files (ADF) are virtual representations of 3.5 inch floppy disks for the Commodore Amiga. Most disk drives from other systems (including modern drives) are not able to read these disks. The 'adfExplorer' package enables you to establish R connections to files on such virtual DOS-formatted disks, which can be use to read from and write to those files.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pepijn-devries.github.io/adfExplorer/}
  \item Report bugs at \url{https://github.com/pepijn-devries/adfExplorer/issues}
}

}
\author{
\strong{Maintainer}: Pepijn de Vries \email{pepijn.devries@outlook.com} (\href{https://orcid.org/0000-0002-7961-6646}{ORCID})

Authors:
\itemize{
  \item Laurent Clévy (Creator of the original ADFlib library) [copyright holder]
}

}
\keyword{internal}
