% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_adf.R
\name{close_all_devices}
\alias{close_all_devices}
\title{Close all virtual devices}
\usage{
close_all_devices()
}
\value{
Returns \code{NULL} invisibly.
}
\description{
Close all virtual devices that are currently open. This function is useful
when you have multiple devices opened at the same time.
}
