% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo.R
\name{demo_adf}
\alias{demo_adf}
\title{Connect with a demonstration ADF file}
\usage{
demo_adf(write_protected = TRUE)
}
\arguments{
\item{write_protected}{A \code{logical} value. When \code{TRUE} you can only read
from the virtual device. If \code{FALSE} the demonstration disk will be
copied to the \code{tempdir()} and you can write to it.}
}
\value{
Returns a connection to a virtual device of class \code{adf_device}
}
\description{
Opens a connection to a virtual device for demonstration purposes
It is used in examples.
}
