% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{amigaDateToRaw}
\alias{amigaDateToRaw}
\alias{amigaIntToRaw}
\alias{bitmapToRaw}
\alias{displayRawData}
\alias{rawToAmigaDate}
\alias{rawToAmigaInt}
\alias{rawToBitmap}
\alias{adf.disk.name<-}
\alias{adf.file.mode<-}
\alias{adf.file.time<-}
\alias{amigaBlock<-}
\alias{current.adf.dir<-}
\alias{adf.disk.name}
\alias{adf.file.exists}
\alias{adf.file.info}
\alias{adf.file.mode}
\alias{adf.file.remove}
\alias{adf.file.size}
\alias{adf.file.time}
\alias{amigaBlock}
\alias{blank.amigaDOSDisk}
\alias{current.adf.dir}
\alias{dir.create.adf}
\alias{dir.exists.adf}
\alias{get.blockID}
\alias{get.diskLocation}
\alias{is.amigaDOS}
\alias{is.bootable}
\alias{list.adf.files}
\alias{put.adf.file}
\alias{read.adf}
\alias{read.adz}
\alias{write.adf}
\alias{write.adz}
\title{Deprecated functions}
\usage{
amigaDateToRaw(...)

amigaIntToRaw(...)

bitmapToRaw(...)

displayRawData(...)

rawToAmigaDate(...)

rawToAmigaInt(...)

rawToBitmap(...)

adf.disk.name(...) <- value

adf.file.mode(...) <- value

adf.file.time(...) <- value

amigaBlock(...) <- value

current.adf.dir(...) <- value

adf.disk.name(...)

adf.file.exists(...)

adf.file.info(...)

adf.file.mode(...)

adf.file.remove(...)

adf.file.size(...)

adf.file.time(...)

amigaBlock(...)

blank.amigaDOSDisk(...)

current.adf.dir(...)

dir.create.adf(...)

dir.exists.adf(...)

get.blockID(...)

get.diskLocation(...)

is.amigaDOS(...)

is.bootable(...)

list.adf.files(...)

put.adf.file(...)

read.adf(...)

read.adz(...)

write.adf(...)

write.adz(...)
}
\arguments{
\item{...}{Ignored.}

\item{value}{Ignored.}
}
\value{
\code{NULL}
}
\description{
Functions documented here are deprecated and will be removed in future
versions of the package. Please use the new functions as indicated instead
or revert to old releases when no alternative is available. See also
\code{vignette("version2")} for more information.
}
