% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02amigaDiskBasics.r, R/03amigaBlockBasics.r
\docType{methods}
\name{print}
\alias{print}
\alias{print,amigaDisk-method}
\alias{print,amigaBlock-method}
\title{Print Amiga Disk File objects}
\usage{
\S4method{print}{amigaDisk}(x, ...)

\S4method{print}{amigaBlock}(x, ...)
}
\arguments{
\item{x}{Either a \code{\link{amigaDisk}} or \code{\link{amigaBlock}} object.}

\item{...}{further arguments passed to or from other methods}
}
\value{
Returns nothing (\code{NULL}).
}
\description{
A method to print Amiga Disk File S4 class objects to the sink.
}
\examples{
data(adf.example)

print(adf.example)
}
\author{
Pepijn de Vries
}
