% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adfExplorer-package.r
\docType{package}
\name{adfExplorer-package}
\alias{adfExplorer}
\alias{adfExplorer-package}
\title{adfExplorer: Import from and Export to Amiga Disk Files}
\description{
Amiga Disk Files (ADF) are virtual representations of 3.5 inch floppy disks for the Commodore Amiga. Most disk drives from other systems (including modern drives) are not able to read these disks. To be able to emulate this system, the ADF format was created. This package enables you to read ADF files and import and export files from and to such virtual DOS-formatted disks.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pepijn-devries/adfExplorer}
  \item \url{https://pepijn-devries.github.io/adfExplorer/}
  \item Report bugs at \url{https://github.com/pepijn-devries/adfExplorer/issues}
}

}
\author{
\strong{Maintainer}: Pepijn de Vries \email{pepijn.devries@outlook.com} [data contributor]

}
\keyword{internal}
